rename background.rs to subscriber.rs
