+++ lib/llm/src/kv_router/subscriber.rs
